Convert an R Markdown driver to one that handles concordances.

These functions produce a new driver which matches the old one in most respects, but adds an argument sourcepos (and possibly others). If that argument is TRUE (the default) then concordances are handled by the new driver.

These functions are used to produce html_documentC and similar drivers in this package, but should also work on other drivers that produce HTML output using Pandoc.

Usage

html_with_concordance(driver)
pdf_with_concordance(driver)

Arguments

driver

An R Markdown driver that produces HTML, LaTeX or PDF using Pandoc.

Value

A new driver function.

Examples

html_with_concordance(rmarkdown::html_fragment)
#> function (sourcepos = TRUE, ...) 
#> {
#>     test_rmarkdown()
#>     res <- driver(...)
#>     res$knitr$opts_knit$concordance <- sourcepos
#>     if (sourcepos) {
#>         res$knitr$opts_knit$out_format <- "markdown"
#>         oldpost <- res$post_processor
#>         res$post_processor <- function(...) {
#>             res <- oldpost(...)
#>             processConcordance(res, res)
#>             res
#>         }
#>         res$pandoc$from <- fix_pandoc_from_options(res$pandoc$from, 
#>             sourcepos)
#>     }
#>     res
#> }
#> <bytecode: 0x55c463bd76d8>
#> <environment: 0x55c463bdca88>
pdf_with_concordance(rmarkdown::latex_fragment)
#> function (latex_engine = "pdflatex", sourcepos = TRUE, defineSconcordance = TRUE, 
#>     run_latex = TRUE, ...) 
#> {
#>     test_rmarkdown()
#>     force(run_latex)
#>     res <- driver(latex_engine = latex_engine, ...)
#>     res$knitr$opts_knit$concordance <- sourcepos
#>     if (sourcepos) {
#>         res$pandoc$from <- fix_pandoc_from_options(res$pandoc$from, 
#>             sourcepos)
#>         res$pandoc$lua_filters <- c(res$pandoc$lua_filters, system.file("rmarkdown/lua/pagebreak.lua", 
#>             package = "patchDVI"), system.file("rmarkdown/lua/latex-datapos.lua", 
#>             package = "patchDVI"))
#>         orig_ext <- res$pandoc$ext
#>         res$pandoc$ext = ".tex"
#>         oldpost <- res$post_processor
#>         res$post_processor <- function(yaml, infile, outfile, 
#>             ...) {
#>             workdir <- dirname(outfile)
#>             concordanceFile <- paste0(sans_ext(normalizePath(infile)), 
#>                 "-concordance.tex")
#>             origdir <- setwd(workdir)
#>             on.exit(setwd(origdir))
#>             processLatexConcordance(outfile, followConcordance = concordanceFile, 
#>                 defineSconcordance = defineSconcordance)
#>             newoutfile <- outfile
#>             if (run_latex && (length(orig_ext) != 1 || orig_ext != 
#>                 ".tex")) {
#>                 consoleLog <- try(texi2dvi(outfile, pdf = TRUE))
#>                 if (!inherits(consoleLog, "try-error")) {
#>                   tempLog <- tempfile(fileext = ".log")
#>                   writeLines(consoleLog, tempLog)
#>                   patchLog(tempLog)
#>                   consoleLog <- readLines(tempLog)
#>                 }
#>                 cat(consoleLog, sep = "\n")
#>                 message(patchSynctex(sub("\\.tex$", ".synctex", 
#>                   outfile, ignore.case = TRUE), patchLog = TRUE))
#>                 newoutfile <- paste0(sans_ext(outfile), ".pdf")
#>             }
#>             if (is.function(oldpost)) 
#>                 res <- oldpost(yaml, infile, newoutfile, ...)
#>             else res <- newoutfile
#>             res
#>         }
#>     }
#>     res
#> }
#> <bytecode: 0x55c463c38d48>
#> <environment: 0x55c463c40450>