R Markdown driver to add concordance

The pdf_documentC driver replaces the like-named rmarkdown driver with one that outputs Commonmark rather than Pandoc Markdown. Commonmark is a dialect of Markdown for which the Pandoc driver for Commonmark supports output of source position information. By using this function as your output driver, you can get that in your own documents.

Usage

pdf_documentC(latex_engine = "pdflatex",
           sourcepos = TRUE,
           defineSconcordance = TRUE,
           run_latex = TRUE, ...)

Arguments

latex_engine

Command to convert .tex file to .pdf.

sourcepos

Whether to include source position information.

defineSconcordance

If TRUE, insert a definition of the \Sconcordance macro just before \begin{document}.

run_latex

Whether to run the latex_engine to produce a PDF.

...

Other arguments to pass to the base driver.

Details

This driver modifies the standard rmarkdown driver pdf_document to use Commonmark and add concordances.

The latex_formatC driver requires a currently unreleased version of the commonmark package, with version higher than 1.9.0.

Value

An R Markdown output format object which will add concordance information.

Author

Duncan Murdoch